refactore(store): Add account store trait
